+++ sdk/keyvault/azsecrets/client.go
@@ -425,10 +425,6 @@ type UpdateSecretPropertiesOptions struct {
 	// placeholder for future optional parameters
 }
 
-func (u UpdateSecretPropertiesOptions) toGenerated() *internal.KeyVaultClientUpdateSecretOptions {
-	return &internal.KeyVaultClientUpdateSecretOptions{}
-}
-
 // UpdateSecretPropertiesResponse contains the underlying response object for the UpdateSecretProperties method
 type UpdateSecretPropertiesResponse struct {
 	Secret
@@ -801,10 +797,10 @@ type ListDeletedSecretsOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ListDeletedSecrets lists all versions of the specified secret. The full secret identifier and attributes are provided
+// NewListDeletedSecretsPager lists all versions of the specified secret. The full secret identifier and attributes are provided
 // in the response. No values are returned for the secrets. This operation requires the secrets/list permission.
-func (c *Client) ListDeletedSecrets(options *ListDeletedSecretsOptions) *runtime.Pager[ListDeletedSecretsResponse] {
-	return runtime.NewPager(runtime.PageProcessor[ListDeletedSecretsResponse]{
+func (c *Client) NewListDeletedSecretsPager(options *ListDeletedSecretsOptions) *runtime.Pager[ListDeletedSecretsResponse] {
+	return runtime.NewPager(runtime.PagingHandler[ListDeletedSecretsResponse]{
 		More: func(page ListDeletedSecretsResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -835,8 +831,8 @@ func (c *Client) ListDeletedSecrets(options *ListDeletedSecretsOptions) *runtime
 	})
 }
 
-// ListSecretVersionsOptions contains the options for the ListSecretVersions operations
-type ListSecretVersionsOptions struct {
+// ListPropertiesOfSecretVersionsOptions contains the options for the ListSecretVersions operations
+type ListPropertiesOfSecretVersionsOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -861,11 +857,11 @@ func listSecretVersionsPageFromGenerated(i internal.KeyVaultClientGetSecretVersi
 	}
 }
 
-// ListPropertiesOfSecretVersions lists all versions of the specified secret. The full secret identifer and
+// NewListPropertiesOfSecretVersionsPager lists all versions of the specified secret. The full secret identifer and
 // attributes are provided in the response. No values are returned for the secrets. This operation
 // requires the secrets/list permission.
-func (c *Client) ListPropertiesOfSecretVersions(name string, options *ListSecretVersionsOptions) *runtime.Pager[ListPropertiesOfSecretVersionsResponse] {
-	return runtime.NewPager(runtime.PageProcessor[ListPropertiesOfSecretVersionsResponse]{
+func (c *Client) NewListPropertiesOfSecretVersionsPager(name string, options *ListPropertiesOfSecretVersionsOptions) *runtime.Pager[ListPropertiesOfSecretVersionsResponse] {
+	return runtime.NewPager(runtime.PagingHandler[ListPropertiesOfSecretVersionsResponse]{
 		More: func(page ListPropertiesOfSecretVersionsResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -896,8 +892,8 @@ func (c *Client) ListPropertiesOfSecretVersions(name string, options *ListSecret
 	})
 }
 
-// ListSecretsOptions contains the options for the ListSecretVersions operations
-type ListSecretsOptions struct {
+// ListPropertiesOfSecretsOptions contains the options for the ListSecretVersions operations
+type ListPropertiesOfSecretsOptions struct {
 	// placeholder for future optional parameters.
 }
 
@@ -922,11 +918,11 @@ func listSecretsPageFromGenerated(i internal.KeyVaultClientGetSecretsResponse) L
 	}
 }
 
-// ListPropertiesOfSecrets list all secrets in a specified key vault. The ListPropertiesOfSecrets operation is applicable to the entire vault,
+// NewListPropertiesOfSecretsPager list all secrets in a specified key vault. The NewListPropertiesOfSecretsPager operation is applicable to the entire vault,
 // however, only the base secret identifier and its attributes are provided in the response. Individual
 // secret versions are not listed in the response. This operation requires the secrets/list permission.
-func (c *Client) ListPropertiesOfSecrets(options *ListSecretsOptions) *runtime.Pager[ListPropertiesOfSecretsResponse] {
-	return runtime.NewPager(runtime.PageProcessor[ListPropertiesOfSecretsResponse]{
+func (c *Client) NewListPropertiesOfSecretsPager(options *ListPropertiesOfSecretsOptions) *runtime.Pager[ListPropertiesOfSecretsResponse] {
+	return runtime.NewPager(runtime.PagingHandler[ListPropertiesOfSecretsResponse]{
 		More: func(page ListPropertiesOfSecretsResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},