+++ sdk/keyvault/azcertificates/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Release History
 
+## 0.4.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 0.4.1 (2022-05-12)
+
+### Other Changes
+* Updated to latest `azcore` and `internal` modules.
+
 ## 0.4.0 (2022-04-21)
 
 ### Breaking Changes